Moved methods from all extended JUnit rules
