+++ src/main/java/com/hotels/beeju/core/BeejuCore.java
@@ -17,18 +17,35 @@ package com.hotels.beeju.core;
 
 import org.apache.derby.jdbc.EmbeddedDriver;
 import org.apache.hadoop.hive.conf.HiveConf;
+import org.apache.hadoop.hive.metastore.HiveMetaStore;
 import org.apache.hadoop.hive.metastore.HiveMetaStoreClient;
 import org.apache.hadoop.hive.metastore.api.Database;
+import org.apache.hadoop.hive.metastore.api.MetaException;
+import org.apache.hadoop.hive.thrift.HadoopThriftAuthBridge;
+import org.apache.hadoop.hive.thrift.HadoopThriftAuthBridge23;
+import org.apache.hive.service.Service;
+import org.apache.hive.service.server.HiveServer2;
 import org.apache.thrift.TException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.io.File;
 import java.io.IOException;
+import java.net.ServerSocket;
 import java.util.Map;
 import java.util.UUID;
+import java.util.concurrent.Callable;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.locks.Condition;
+import java.util.concurrent.locks.Lock;
+import java.util.concurrent.locks.ReentrantLock;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
 public class BeejuCore {
+  private static final Logger LOG = LoggerFactory.getLogger(BeejuCore.class);
 
   // "user" conflicts with USER db and the metastore_db can't be created.
   public static final String METASTORE_DB_USER = "db_user";
@@ -38,6 +55,7 @@ public class BeejuCore {
   private final String databaseName;
   private final String connectionURL;
   private final String driverClassName;
+  private int thriftPort;
 
   public BeejuCore(String databaseName, Map<String, String> configuration){
     checkNotNull(databaseName, "databaseName is required");
@@ -122,6 +140,99 @@ public class BeejuCore {
    */
   public String connectionURL() {
     return connectionURL;
+  }
+
+  public void startThrift(ExecutorService thriftServer) throws Exception {
+    thriftPort = -1;
+    final Lock startLock = new ReentrantLock();
+    final Condition startCondition = startLock.newCondition();
+    final AtomicBoolean startedServing = new AtomicBoolean();
+    try (ServerSocket socket = new ServerSocket(0)) {
+      thriftPort = socket.getLocalPort();
+    }
+    setHiveVar(HiveConf.ConfVars.METASTOREURIS, getThriftConnectionUri());
+    final HiveConf hiveConf = new HiveConf(conf(), HiveMetaStoreClient.class);
+    thriftServer.execute(new Runnable() {
+      @Override
+      public void run() {
+        try {
+          HadoopThriftAuthBridge bridge = new HadoopThriftAuthBridge23();
+          HiveMetaStore.startMetaStore(thriftPort, bridge, hiveConf, startLock, startCondition, startedServing);
+        } catch (Throwable e) {
+          LOG.error("Unable to start a Thrift server for Hive Metastore", e);
+        }
+      }
+    });
+    int i = 0;
+    while (i++ < 3) {
+      startLock.lock();
+      try {
+        if (startCondition.await(1, TimeUnit.MINUTES)) {
+          break;
+        }
+      } finally {
+        startLock.unlock();
+      }
+      if (i == 3) {
+        throw new RuntimeException("Maximum number of tries reached whilst waiting for Thrift server to be ready");
+      }
+    }
+  }
+
+  /**
+   * @return The Thrift connection {@link URI} string for the Metastore service.
+   */
+  public String getThriftConnectionUri() {
+    return "thrift://localhost:" + thriftPort;
+  }
+
+  /**
+   * @return The port used for the Thrift Metastore service.
+   */
+  public int getThriftPort() {
+    return thriftPort;
+  }
+
+  public void waitForHiveServer2StartUp(HiveServer2 hiveServer2) throws InterruptedException {
+    int retries = 0;
+    int maxRetries = 5;
+    while (hiveServer2.getServiceState() != Service.STATE.STARTED && retries < maxRetries) {
+      Thread.sleep(1000);
+      retries++;
+    }
+    if (retries >= maxRetries) {
+      throw new RuntimeException("HiveServer2 did not start in a reasonable time");
+    }
+  }
+
+  /**
+   * Creates a new HiveMetaStoreClient that can talk directly to the backed metastore database.
+   * <p>
+   * The invoker is responsible for closing the client.
+   * </p>
+   *
+   * @return the {@link HiveMetaStoreClient} backed by an HSQLDB in-memory database.
+   */
+  public HiveMetaStoreClient newClient() {
+    try {
+      return new HiveMetaStoreClient(conf());
+    } catch (MetaException e) {
+      throw new RuntimeException("Unable to create HiveMetaStoreClient", e);
+    }
+  }
+
+  public static class CallableHiveClient implements Callable<HiveMetaStoreClient> {
+
+    private final HiveConf hiveConf;
+
+    public CallableHiveClient(HiveConf hiveConf) {
+      this.hiveConf = hiveConf;
+    }
+
+    @Override
+    public HiveMetaStoreClient call() throws Exception {
+      return new HiveMetaStoreClient(hiveConf);
+    }
   }
 
 }
